Mudança na lógica das carteiras e implementação das transações com refatoramento de algumas classes

+++ src/main/java/com/picpay_challenge/dto/PersonDto.java
@@ -3,6 +3,7 @@ package com.picpay_challenge.dto;
 import com.picpay_challenge.enums.AccountType;
 import lombok.NonNull;
 
+import java.math.BigDecimal;
 
 
 public record PersonDto(@NonNull String name,
@@ -10,4 +11,4 @@ public record PersonDto(@NonNull String name,
                         @NonNull String password,
                         @NonNull String document,
                         @NonNull AccountType accountType,
-                        double balance){};
+                        BigDecimal balance){};
